fix stop start toggle not beeing applied on creation; MoreOptionsDefaults is now a separate class
